+++ typescript/tests.ts
@@ -530,7 +530,7 @@ isRecoilValue(mySelector1);
         setSelf('a'); // $ExpectError
 
         onSet(val => {
-          val; // $ExpectType number | DefaultValue
+          val; // $ExpectType number
         });
         onSet('a'); // $ExpectError
 
@@ -556,7 +556,7 @@ isRecoilValue(mySelector1);
         setSelf('a'); // $ExpectError
 
         onSet(val => {
-          val; // $ExpectType number | DefaultValue
+          val; // $ExpectType number
         });
         onSet('a'); // $ExpectError